Maintenance/update-assembly-version (#8)
* Add GitVersion step to update AssemblyInfo on 'main'

Modified `Build.yml` to include a new step for updating the
AssemblyInfo file using `GitVersion.Tool`. This step installs
the tool globally and runs it to update assembly information
before the build process. The step is conditional and only
executes if the current branch is 'main'.

* Remove old OpenTelemetry.TraceListener folders.

These changes indicate a complete removal of the `OpenTelemetry.TraceListener` project and its associated test project, suggesting a significant restructuring or deprecation.

* Add updateAssemblyInfo param; remove Set Outputs step

Added the `updateAssemblyInfo` parameter to the `version_step` job in `Build.yml`, setting it to `true`. Removed the "📝 Set Outputs" step, which previously echoed the `fullSemVer` output from the `version_step` job to the GitHub Actions output.

* Refactor CI: Merge version calc into build job

Removed the "Calculate Version" job from Build.yml and integrated its steps into the "Build and Test" job. This includes installing GitVersion, determining the version, and setting the version outputs. Additionally, changed fetch-depth for checkout action from 1 to 0 to fetch full repository history.

* Refactor NuGet package versioning in build workflow

* Refactor build workflow to update AssemblyInfo on 'main' branch

* Refactor build workflow to remove 'maintenance/update-assembly-version' branch from trigger